chore: minor refactor for better types
