+++ packages/auth-provider/src/common/utilities.ts
@@ -1,60 +1,31 @@
 import {
 	API_TYPE,
 	AUTH_TYPES,
-	HEADERS,
 	JWT,
 	decodeToken,
 	verifyAndExtractToken,
 } from "@versini/auth-common";
 import { getFingerprintHash } from "@versini/ui-fingerprint";
-
-import { API_ENDPOINT } from "./constants";
-import type { ServiceCallProps } from "./types";
+import { restCall } from "./services";
+import type {
+	AuthenticateUserProps,
+	AuthenticateUserResponse,
+	BooleanResponse,
+	GetAccessTokenSilentlyProps,
+	GetAccessTokenSilentlyResponse,
+	GetPreAuthCodeProps,
+	LogoutProps,
+} from "./types";
 
 const isProd = process.env.NODE_ENV === "production";
-const isDev = !isProd;
+export const isDev = !isProd;
 
-export const getUserIdFromToken = (token: string) => {
+export const getUserIdFromToken = (token: string): string => {
 	try {
 		const jwt = decodeToken(token);
 		return jwt ? (jwt[JWT.USER_ID_KEY] as string) : "";
 	} catch (_error) {
 		return "";
-	}
-};
-
-export const serviceCall = async ({
-	type,
-	clientId,
-	params = {},
-}: ServiceCallProps) => {
-	try {
-		const response = await fetch(
-			isDev ? `${API_ENDPOINT.dev}/${type}` : `${API_ENDPOINT.prod}/${type}`,
-			{
-				credentials: "include",
-				method: "POST",
-				headers: {
-					"Content-Type": "application/json",
-					[HEADERS.CLIENT_ID]: `${clientId}`,
-				},
-				body: JSON.stringify(params),
-			},
-		);
-
-		if (response.status !== 200) {
-			return { status: response.status, data: [] };
-		}
-		const { data, errors } = await response.json();
-
-		return {
-			status: response.status,
-			data,
-			errors,
-		};
-	} catch (_error) {
-		console.error(_error);
-		return { status: 500, data: [] };
 	}
 };
 
@@ -65,16 +36,9 @@ export const logoutUser = async ({
 	refreshToken,
 	clientId,
 	domain,
-}: {
-	userId: string;
-	idToken: string;
-	accessToken: string;
-	refreshToken: string;
-	clientId: string;
-	domain: string;
-}) => {
+}: LogoutProps): Promise<BooleanResponse> => {
 	try {
-		const response = await serviceCall({
+		const response = await restCall({
 			type: API_TYPE.LOGOUT,
 			clientId,
 			params: {
@@ -95,18 +59,6 @@ export const logoutUser = async ({
 	}
 };
 
-export type AuthenticateUserProps = {
-	username: string;
-	password: string;
-	clientId: string;
-	nonce: string;
-	type?: string;
-	sessionExpiration?: string;
-	code?: string;
-	code_verifier?: string;
-	domain: string;
-	fingerprint: string;
-};
 export const authenticateUser = async ({
 	username,
 	password,
@@ -118,9 +70,9 @@ export const authenticateUser = async ({
 	code_verifier,
 	domain,
 	fingerprint,
-}: AuthenticateUserProps) => {
+}: AuthenticateUserProps): Promise<AuthenticateUserResponse> => {
 	try {
-		const response = await serviceCall({
+		const response = await restCall({
 			type: API_TYPE.AUTHENTICATE,
 			clientId,
 			params: {
@@ -164,9 +116,9 @@ export const getPreAuthCode = async ({
 	nonce,
 	clientId,
 	code_challenge,
-}: { clientId: string; nonce: string; code_challenge: string }) => {
+}: GetPreAuthCodeProps) => {
 	try {
-		const response = await serviceCall({
+		const response = await restCall({
 			type: API_TYPE.CODE,
 			clientId,
 			params: {
@@ -192,14 +144,6 @@ export const getPreAuthCode = async ({
 	}
 };
 
-export type getAccessTokenSilently = {
-	clientId: string;
-	userId: string;
-	nonce: string;
-	refreshToken: string;
-	accessToken: string;
-	domain: string;
-};
 export const getAccessTokenSilently = async ({
 	clientId,
 	userId,
@@ -207,9 +151,9 @@ export const getAccessTokenSilently = async ({
 	refreshToken,
 	accessToken,
 	domain,
-}: getAccessTokenSilently) => {
+}: GetAccessTokenSilentlyProps): Promise<GetAccessTokenSilentlyResponse> => {
 	try {
-		const response = await serviceCall({
+		const response = await restCall({
 			type: API_TYPE.AUTHENTICATE,
 			clientId,
 			params: {
@@ -246,151 +190,7 @@ export const getAccessTokenSilently = async ({
 	}
 };
 
-const GRAPHQL_QUERIES = {
-	GET_REGISTRATION_OPTIONS: `mutation GetPasskeyRegistrationOptions(
-	$clientId: String!,
-	$username: String!,
-	$id: String!) {
-		getPasskeyRegistrationOptions(clientId: $clientId, username: $username, id: $id) {
-			challenge
-			rp {
-				id
-				name
-			}
-			user {
-				id
-				name
-				displayName
-			}
-			pubKeyCredParams {
-				type
-				alg
-			}
-			timeout
-			attestation
-		}
-	}`,
-	VERIFY_REGISTRATION: `mutation VerifyPasskeyRegistration(
-		$clientId: String!,
-		$username: String!,
-		$id: String!,
-		$registration: RegistrationOptionsInput!) {
-		verifyPasskeyRegistration(
-			clientId: $clientId,
-			username: $username,
-			id: $id,
-			registration: $registration) {
-			status
-			message
-		}
-	}`,
-	GET_AUTHENTICATION_OPTIONS: `mutation GetPasskeyAuthenticationOptions(
-		$id: String!,
-		$clientId: String!,
-		) {
-		getPasskeyAuthenticationOptions(
-			id: $id,
-			clientId: $clientId) {
-				rpId,
-				challenge,
-				allowCredentials {
-					id,
-					type,
-					transports
-				}
-				timeout,
-				userVerification,
-		}
-	}`,
-	VERIFY_AUTHENTICATION: `mutation VerifyPasskeyAuthentication(
-		$clientId: String!,
-		$id: String!,
-		$authentication: AuthenticationOptionsInput!,
-		$nonce: String!,
-		$domain: String,
-		$fingerprint: String) {
-		verifyPasskeyAuthentication(
-			clientId: $clientId,
-			id: $id,
-			authentication: $authentication,
-			nonce: $nonce,
-			domain: $domain,
-			fingerprint: $fingerprint) {
-				status,
-				idToken,
-				accessToken,
-				refreshToken,
-				userId,
-				username,
-		}
-	}`,
-};
-export const SERVICE_TYPES = {
-	GET_REGISTRATION_OPTIONS: {
-		schema: GRAPHQL_QUERIES.GET_REGISTRATION_OPTIONS,
-		method: "getPasskeyRegistrationOptions",
-	},
-	VERIFY_REGISTRATION: {
-		schema: GRAPHQL_QUERIES.VERIFY_REGISTRATION,
-		method: "verifyPasskeyRegistration",
-	},
-	GET_AUTHENTICATION_OPTIONS: {
-		schema: GRAPHQL_QUERIES.GET_AUTHENTICATION_OPTIONS,
-		method: "getPasskeyAuthenticationOptions",
-	},
-	VERIFY_AUTHENTICATION: {
-		schema: GRAPHQL_QUERIES.VERIFY_AUTHENTICATION,
-		method: "verifyPasskeyAuthentication",
-	},
-};
-
-export const graphQLCall = async ({
-	accessToken,
-	type,
-	clientId,
-	params = {},
-}: {
-	accessToken: string;
-	clientId: string;
-	type: any;
-	params?: any;
-}) => {
-	try {
-		const requestData = type?.data ? type.data(params) : params;
-		const authorization = `Bearer ${accessToken}`;
-		const response = await fetch(
-			isDev ? `${API_ENDPOINT.dev}/graphql` : `${API_ENDPOINT.prod}/graphql`,
-			{
-				method: "POST",
-				credentials: "include",
-				headers: {
-					authorization,
-					"Content-Type": "application/json",
-					Accept: "application/json",
-					[HEADERS.CLIENT_ID]: `${clientId}`,
-				},
-				body: JSON.stringify({
-					query: type.schema,
-					variables: requestData,
-				}),
-			},
-		);
-		if (response.status !== 200) {
-			return { status: response.status, data: [] };
-		}
-		const { data, errors } = await response.json();
-		return {
-			status: response.status,
-			data: data[type.method],
-			errors,
-		};
-	} catch (_error) {
-		console.error(_error);
-		return { status: 500, data: [] };
-	}
-};
-
-export const getCustomFingerprint = async () => {
+export const getCustomFingerprint = async (): Promise<string> => {
 	try {
 		return await getFingerprintHash();
 	} catch (_error) {
